+++ app/src/main/java/com/example/wiphonepe/data/FilterData.kt
@@ -2,6 +2,6 @@ package com.example.wiphonepe.data
 
 data class FilterData(
     val sameDayShipping: Boolean = false,
-    val minPrice: Double = 0.0,
-    val maxPrice: Double = 1e4
+    val minPrice: Float = 0f,
+    val maxPrice: Float = 1e4f
 )
